requirements.txt
asgiref==3.8.1
bareun-apis==0.12.0
bareunpy==1.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
Django==5.1.2
djangorestframework==3.15.2
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
googleapis-common-protos==1.65.0
grpcio==1.66.2
huggingface-hub==0.25.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
JPype1==1.5.0
kiwisolver==1.4.7
konlpy==0.6.0
lxml==5.3.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.4.1
nltk==3.9.1
numpy==2.1.2
packaging==24.1
pillow==10.4.0
protobuf==3.20.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
urllib3==2.2.3
wordcloud==1.9.3
